Renaming to CorsSupport to ...Base and CorsSupportSe to CorsSupport; further clean-up of builder/config/'from'
