+++ webserver/cors/src/test/java/io/helidon/webserver/cors/CrossOriginConfigTest.java
@@ -49,7 +49,7 @@ public class CrossOriginConfigTest {
         Config node = testConfig.get("narrow");
         assertThat(node, is(notNullValue()));
         assertThat(node.exists(), is(true));
-        CrossOriginConfig c = node.as(CrossOriginConfig::build).get();
+        CrossOriginConfig c = node.as(CrossOriginConfig::create).get();
 
         assertThat(c.isEnabled(), is(true));
         assertThat(c.allowOrigins(), arrayContaining("http://foo.bar", "http://bar.foo"));
@@ -63,7 +63,7 @@ public class CrossOriginConfigTest {
     @Test
     public void testMissing() {
         Assertions.assertThrows(MissingValueException.class, () -> {
-                CrossOriginConfig basic = testConfig.get("notThere").as(CrossOriginConfig::build).get();
+                CrossOriginConfig basic = testConfig.get("notThere").as(CrossOriginConfig::create).get();
         });
     }
 
@@ -72,7 +72,7 @@ public class CrossOriginConfigTest {
         Config node = testConfig.get("wide");
         assertThat(node, is(notNullValue()));
         assertThat(node.exists(), is(true));
-        CrossOriginConfig b = node.as(CrossOriginConfig::build).get();
+        CrossOriginConfig b = node.as(CrossOriginConfig::create).get();
 
         assertThat(b.isEnabled(), is(false));
         assertThat(b.allowOrigins(), arrayContaining(ALLOW_ALL));
@@ -88,7 +88,7 @@ public class CrossOriginConfigTest {
         Config node = testConfig.get("just-disabled");
         assertThat(node, is(notNullValue()));
         assertThat(node.exists(), is(true));
-        CrossOriginConfig b = node.as(CrossOriginConfig::build).get();
+        CrossOriginConfig b = node.as(CrossOriginConfig::create).get();
 
         assertThat(b.isEnabled(), is(false));
     }
@@ -98,7 +98,7 @@ public class CrossOriginConfigTest {
         Config node = testConfig.get("cors-setup");
         assertThat(node, is(notNullValue()));
         assertThat(node.exists(), is(true));
-        MappedCrossOriginConfig m = node.as(MappedCrossOriginConfig::from).get();
+        MappedCrossOriginConfig m = node.as(MappedCrossOriginConfig::create).get();
 
         assertThat(m.isEnabled(), is(true));
